Use longer timeout values for AC checks and CAS access.
The v0.0.15 tag (commit 1933b8e, "Import changes for goma server")
introduced two changes that impact the efficiency of the Goma server:

- 206814d1c610c, "remoteexec: set short timeout for cas download"
- 86cdb70a44908, "remoteexec: set short timeout in each stage in Exec
                  request"

Which introduced a 1 second timeout value for CAS access and a 1 second
timeout value for checking the AC for cached actions, respectively.

1 second can be too little depending on the backend and the infrastructure
setup (curiously, increasing the value to as little as 2 seconds is enough
to have each of those two actions complete in a few miliseconds).

Be more lenient and use a 3-second timeout instead.

Change-Id: I755dd6db7cbb14a1233b4715a607bef07c08b9e2